+++ CHANGES.rst
@@ -5,6 +5,9 @@ Version 2.2.0
 
 Unreleased
 
+-   Add an ``app.redirect`` method, which ``flask.redirect`` will call.
+    This makes it possible for an app to override how redirects work.
+    :issue:`4569`
 -   Refactor ``register_error_handler`` to consolidate error checking.
     Rewrite some error messages to be more consistent. :issue:`4559`